Move debug-tests util functions into main tests utils
